requirements.txt
Django>5.2,<5.3
djangorestframework>3.16,<3.17